Android: Ability to easily create tests for vanilla android modules
